requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.0
bs4==0.0.1
cachetools==2.0.1
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
click==6.7
comtypes==1.1.4
configparser==3.7.3
cryptography==2.6.1
dill==0.2.7.1
enum34==1.1.6
ffmpy==0.2.2
Flask==0.12.2
Flask-Cors==3.0.3
future==0.16.0
google==2.0.2
google-api-core==1.8.2
google-api-python-client==1.7.8
google-auth==1.3.0
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-core==0.28.0
google-cloud-language==1.2.0
google-cloud-speech==1.0.0
google-cloud-texttospeech==0.4.0
google-cloud-vision==0.36.0
google-gax==0.15.16
googleapis-common-protos==1.5.3
grpcio==1.19.0
httplib2==0.12.1
idna==2.6
ipaddress==1.0.19
itsdangerous==0.24
Jinja2==2.10
language==0.6
MarkupSafe==1.0
ply==3.10
protobuf==3.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pyOpenSSL==17.5.0
pyreadline==2.1
pytz==2017.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
soupsieve==1.8
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.14.1
